Simplifies get_top_bits.
Previously, a subtle bug manifested itself due
to our logic assuming that when a mask (1s)
aligned to MSB of value will never be greater
than that value until mask shifted once more to the
left. This assumption led to failed test case for
value = 42069 and num_top_bits = 4.

We now shift ~0 (a mask of all 1s) to the left
until the last 1 in the mask aligns with the MSB of
value and is succeeded by num_top_bit - 1 number of ones.

After we bit-wise AND the mask and value, we truncate the
remaining bits by shifting right.