Generic Queue
This is a circular linked list serving
as a queue with O(1) enqueueing and dequeueing.

Currently, the node's scheme indicates data is an int.
When porting this to project 2, data will take on
the value of a tcb pointer.

Further enqueueing and dequeueing + prints are tested
to play around with.