Mutex Logic
I have fleshed out the logic for mutexes in code.
See extensive comments within lock and unlock function.

Note that since the mutexes list is a shared resource,
we must synchronize access to the shared resource
to avoid concurrency bugs. Therefore, the block signals
and unblock signals are necessary when entering and exiting
functions.

The mutex_node and mutex_list do not merit their own file, so
they can be ported to thread_worker.c. The mutex_list and
supporting mutex_node is necessary for init and destroy mutex
functions. We must keep track of them.

See also the very important attribute added to tcb type.