Mutexes Overhauled
Because a worker does not need to know which thread
holds a particular lock, rather, whether the worker itself
holds the lock -> we can simply our worker_mutex_t type.

By doing so, we must shift where we keep track of which thread
holds which lock. Note that this is important for us because
the scheduler skips the thread waiting on a mutex. While this
could've been simpler if we allow a thread to be scheduled despite
waiting on another thread/lock, we choose to incorporate a little
extra state to prevent too many context switches (scheduler logic
skips those).

This overhaul helps with the new changes. Note that mutex_node
are kept in the heap along with the mutexes list.