A more fair scheduler.
Whenever a thread needs to be scheduled and all other threads are
blocked, there is no reason to set a timer on the to be scheduled
thread. By doing so, we are context switching out of the thread
for no reason because the other threads won't be scheduled anyway.

The only exception to this is if the thread unlocks a mutex upon
which one or more other threads were blocked. For the scheduled
thread to have not been given a timer and unlocked a mutex, we must
swap context to the scheduler to allow the other threads the chance
to acquire that mutex. If we do not, the same scheduled thread can
later acquire the mutex.

In short, when a thread is no longer blocked, we must give it
the chance to be scheduled.