Reduced Context Switches
By yield each time we unlock a mutex, the context switch #s are:
- parallel_cal = 20,740
- vector_mult = 6 million
- external cal = 20 million

By yielding unlock only if one other thread unblocked as a result
of releasing the mutex:
- parallel calc = 500
- vector_mult = 2700
- external calc = 9610

By setting a timer for a thread only if remaining threads are
waiting to join as opposted to setting a timer iff remaining threads
are either blocked waiting on another thread or waiting for a mutex,
we get reduced numbers again. Note that in this change, we set the
timer if some other thread is waiting on the mutex. This change allows
us to simplify unlock logic to remove the abstraction leak about
the scheduler. By doing so, the thread currently running will be
preempted and a context switch will happen less often than had we
yielded upon finding one other thread which no longer waiting on mutex.
- parallel _ cal = 480
- vector_multipl = 2020
- external_calc  = 7232