Fix assumption of quantum_amt_used
Previously, we assumed that if the worker only used a slice
of the quantum, (a portion of the whole), that adding and having
the sum going over the TIME_QUANTUM amt wouldn't be that big of a
problem. This assumption fails with PSJF and atomics because the
quantums_elapsed is never booted down so the same thread yields()
continuously.