Second (#13)
* p2 files

* Creates struct and starts worker_create

Adds recommended variables in the struct.
Might need more changes in future based on what is needed.
Lines in worker_create uses the struct created to allocate memory / initialize variables.
STACK_SIZE constant is arbitrary need to look into it.

* Revert "Creates struct and starts worker_create"

This reverts commit 0ba9c2a4d8c4bd093136d71a5b1e9c9b7156c2c1.

* Additional Example Files

* Generic Queue

This is a circular linked list serving
as a queue with O(1) enqueueing and dequeueing.

Currently, the node's scheme indicates data is an int.
When porting this to project 2, data will take on
the value of a tcb pointer.

Further enqueueing and dequeueing + prints are tested
to play around with.

* Porting code for part1.

The coding for this part has been done in different
branches. I have ported those changes here.

The reason for doing them in other branches was to have
all code (prototypes, defs) in one tw.c file. VSCode allowed
easy debugging from within one file.

* Refactored join

* updated makefiles

* make works now

* Total Context Switches Counting

* A more fair scheduler.

Whenever a thread needs to be scheduled and all other threads are
blocked, there is no reason to set a timer on the to be scheduled
thread. By doing so, we are context switching out of the thread
for no reason because the other threads won't be scheduled anyway.

The only exception to this is if the thread unlocks a mutex upon
which one or more other threads were blocked. For the scheduled
thread to have not been given a timer and unlocked a mutex, we must
swap context to the scheduler to allow the other threads the chance
to acquire that mutex. If we do not, the same scheduled thread can
later acquire the mutex.

In short, when a thread is no longer blocked, we must give it
the chance to be scheduled.

* Benchmarking mechanics

* STACK_SIZE instead of 4096

* Revert "STACK_SIZE instead of 4096"

This reverts commit 2f14bf65aebff6fe005740d2e4bd243b48797357.

* Remove some printlns

* Timer fix (#6)

* Resetting timer

* Timer appropriately set

* Uses defined SCHED to call right scheduler

* Worker_join refactored

* changes back to original file

* Achieves Context Switch Reduction

This change incorporates one refactoring and one scheduling logic
change. When a thread gives up a lock, the scheduler - if it had
not put a timer on that thread bc other threads were blocked - does
not get passed control if the unlock did not unblock any other thread.

* Reduced Context Switches

By yield each time we unlock a mutex, the context switch #s are:
- parallel_cal = 20,740
- vector_mult = 6 million
- external cal = 20 million

By yielding unlock only if one other thread unblocked as a result
of releasing the mutex:
- parallel calc = 500
- vector_mult = 2700
- external calc = 9610

By setting a timer for a thread only if remaining threads are
waiting to join as opposted to setting a timer iff remaining threads
are either blocked waiting on another thread or waiting for a mutex,
we get reduced numbers again. Note that in this change, we set the
timer if some other thread is waiting on the mutex. This change allows
us to simplify unlock logic to remove the abstraction leak about
the scheduler. By doing so, the thread currently running will be
preempted and a context switch will happen less often than had we
yielded upon finding one other thread which no longer waiting on mutex.
- parallel _ cal = 480
- vector_multipl = 2020
- external_calc  = 7232

* Revert "changes back to original file"

This reverts commit c5204191c9d237f168f7da335ed1355eae37522e.

* Increased stack size to fix unexpected bug

* fix unnecessary revert

* Changes time quantum to milliseconds

* Psjf (#7)

* WIP 1

* PSJF Without Formal testing of insert

* Remove unnecessary sync mechanism

* Revert "Remove unnecessary sync mechanism"

This reverts commit 41a0ae10c8b6b1ced5cb4e15ba4d5d5ce5b59bfc.

* Fix assumption of quantum_amt_used

Previously, we assumed that if the worker only used a slice
of the quantum, (a portion of the whole), that adding and having
the sum going over the TIME_QUANTUM amt wouldn't be that big of a
problem. This assumption fails with PSJF and atomics because the
quantums_elapsed is never booted down so the same thread yields()
continuously.

* Remove time_quantum_used_up_fully from PSJF (#8)

This flag is unnecessary in PSJF since the insertion into tcbs
list happens based on quantums elapsed and quantum time elapsed
anyway.

* Mlfq (#9)

* WIP

* WIP2 does not work correctly

* removes prints

* MLFQ with Renewed boosting algorithm

* Calculates microseconds correctly

---------

Co-authored-by: Yash <58752969+imyash01@users.noreply.github.com>

* Simplifies Codebase (#10)

* Remove unused BOOST_TIME definition

* Remove unused time struct

* Change m.i. counters to be atomic (and not on heap)

* Refactor worker_exit()

Cleanup context alerts threads in case worker_exit() didn't call
it. However, this was redundant.

* ifndef Macro fix (#12)

* Insert function fix

* Submission Files

---------

Co-authored-by: Yash <58752969+imyash01@users.noreply.github.com>