Translating Virtual Addr to Physical Addr
There are three steps necessary in translating a va to pa in
our virtual memory project.

- The first is to accept a virtual address and parse it. Ex: top
10 bits for root pgdir, next 10 bits for intermediate page table,
and finally 12 bits for offset.

- The second step relates to answering whether a va points to a
frame which is currently allocated (or if it is free).

- The third step is to derive the frame from the va and return a
(void *) by incrementing the frame pointer by the byte offset
derived as a result of parsing the offset bits in the virtual
address.

This commit introduces a framework for accomplishing this smoothly.
Testing remains.