Valid Pages Linked Functions
If the user frees less than they malloc, we allow it. They are
responsible for not causing memory leaks. This discrepancy is due
to the free(va, size) method signature which is arbitrary in
whether we have to enforce size to be equivalent to what they
malloc-d(size).

Additionally, this change allows the user library to call
put_value and write less than they malloc-ed which is perfectly
fine.